nput text fields are single line entry fields, which is probably required by home assistant. There must be a standard way to add lists of fields. I would presume one line default, then another field can be added for the second one, in addition a line can be removed with a delete button. Please implement that for both media_path and exclude pattern.
for the selected_image sensor, add the attribute : entity_picture and set it as well to the selected image file.

The paths on the file system are different than what should be set for the selected_picture. While the search path starts with /media/, this needs to be r eplaced image to be displayed as:

media-source://media_source/local/

Add two configuration variables: one with file system prefix, and display uri prefix into the config flow. set file_system_prefix per default to /media, set display_uri_prefix to media-source://media_source/local/, but the user can change these in the config flow. At run time perform the prefix substitution.